M6T16 created smart view component
